scripts/requirements.txt
fastapi
uvicorn
mailchimp_marketing
python-dotenv
requests
python-multipart
